# src/Docs/swagger_documentation.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is vendiman backend API here you can purchase your favourite food products",
        "version": "1.0.6",
        "title": "vendiman",
        "contact": { "email": "pluralisghtvendiman@gmail.com" }
    },
    "host": "localhost:8080",
    "basePath": "/",
...
    "schemes": ["http"],
    "paths": {
...
        "/inventory/items": {
            "post": {
                "tags": ["inventory"],
                "summary": "Post items on particular inventory",
                "description": "need to post with correct items_ids and inventory_id",
                "operationId": "postInventoryItems",
                "produces": ["application/json", "application/xml"],
                "parameters": [
                    {
                        "name": "Inventory Items",
                        "in": "body",
                        "description": "Need to post correct inventory_id and item_id and qty",
                        "required": true,
                        "type": "object",
                        "schema": { "$ref": "#/definitions/Inventory Items" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully posted items in inventory"
                    },
                    "400": { "description": "Invalid status value" }
                },
                "security": [{ "Bearer": ["write:items", "read:items"] }]
            },
            "put": {
                "tags": ["inventory"],
                "summary": "Post items on particular inventory",
                "description": "need to post with correct items_ids and inventory_id",
                "operationId": "putInventoryItems",
                "produces": ["application/json", "application/xml"],
                "parameters": [
                    {
                        "name": "Inventory Items",
                        "in": "body",
                        "description": "Need to post correct inventory_id and item_id and qty",
                        "required": true,
                        "type": "object",
                        "schema": { "$ref": "#/definitions/Inventory Items" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully posted items in inventory"
                    },
                    "400": { "description": "Invalid status value" }
                },
                "security": [{ "Bearer": ["write:items", "read:items"] }]
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "description": "Enter the token with the `Bearer: ` prefix, e.g. Bearer abcde12345"
        },
        "inventory_id": {
            "type": "apiKey",
            "name": "inventory_id",
            "in": "header",
            "description": "Enter valid inventory id"
        }
    },
...
}